pyproject.toml
[project]
name = "gitrep-scorer"
version = "0.1.0"
description = "A program that helps you give a score to a github repository using TOPSIS analysis and an extensive dataset."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair==4.2.0",
    "attrs==22.1.0",
    "beautifulsoup4==4.11.1",
    "blinker==1.5",
    "bs4==0.0.1",
    "cachetools==5.2.0",
    "certifi>=2023.0.0",
    "charset-normalizer==2.1.1",
    "click==8.1.3",
    "colorama==0.4.6",
    "commonmark==0.9.1",
    "decorator==5.1.1",
    "entrypoints==0.4",
    "gitdb==4.0.10",
    "gitpython==3.1.29",
    "html5lib==1.1",
    "idna==3.4",
    "importlib-metadata==5.1.0",
    "importlib-resources==5.10.1",
    "jinja2==3.1.2",
    "jsonschema==4.17.3",
    "markupsafe==2.1.1",
    "numpy>=1.24.0",
    "packaging==22.0",
    "pandas>=2.0.0",
    "pillow==9.3.0",
    "pkgutil-resolve-name==1.3.10",
    "protobuf==3.20.3",
    "pyarrow>=14.0.0",
    "pydeck==0.8.0",
    "pygments==2.13.0",
    "pympler==1.0.1",
    "pyrsistent==0.19.2",
    "python-dateutil==2.8.2",
    "pytz==2022.6",
    "pytz-deprecation-shim==0.1.0.post0",
    "requests==2.28.1",
    "rich==12.6.0",
    "semver==2.13.0",
    "six==1.16.0",
    "smmap==5.0.0",
    "soupsieve==2.3.2.post1",
    "streamlit>=1.28.0",
    "toml==0.10.2",
    "toolz==0.12.0",
    "tornado==6.2",
    "typing-extensions>=4.8.0",
    "tzdata==2022.7",
    "tzlocal==4.2",
    "urllib3==1.26.13",
    "validators==0.20.0",
    "watchdog==2.2.0",
    "webencodings==0.5.1",
    "zipp==3.11.0",
]
